Rename StartOperation to nextOp
Renamed StartOperation to nextOp to better reflect its purpose of
transitioning to the next stage, whether that is within a given
instruction, or to the next instruction. Also eliminated FinishOperation
and integrated its functionality into nextOp. Removed finishAfterWrite
in favor of directly setting m_action to nextOp after write operations.