Rewrote microcode pump to improve cycle accuracy
Rewrote the microcode pump, to read from the bus in the middle of a
microcode step, rather than at the end of a step. This allows a much
easier implementation without needing to request a bus read on one
cycle and store the results on the next cycle. All of the Harte, random
tests pass except for add and subtract since they depend on decimal mode
which is not yet implemented.

The old tests have been disabled. I am considering just deleting the old
tests since I am concerned they are not correct.

Disassembly is broken as well, but will probably be my next task.